# internal/db/generated/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package generated

import (
	"encoding/json"
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type ApiKey struct {
	ID             uuid.UUID          `json:"id"`
	OrganizationID uuid.UUID          `json:"organization_id"`
	Name           string             `json:"name"`
	KeyHash        string             `json:"key_hash"`
	KeyPrefix      string             `json:"key_prefix"`
	CreatedAt      time.Time          `json:"created_at"`
	RevokedAt      pgtype.Timestamptz `json:"revoked_at"`
}

type AuditLog struct {
	ID             uuid.UUID       `json:"id"`
	OrganizationID uuid.UUID       `json:"organization_id"`
	SourceID       uuid.UUID       `json:"source_id"`
	InputHash      pgtype.Text     `json:"input_hash"`
	GuardResults   json.RawMessage `json:"guard_results"`
	OverallResult  string          `json:"overall_result"`
	LatencyMs      int32           `json:"latency_ms"`
	CreatedAt      time.Time       `json:"created_at"`
}

type Destination struct {
	ID        uuid.UUID       `json:"id"`
	SourceID  uuid.UUID       `json:"source_id"`
	Name      string          `json:"name"`
	GuardType string          `json:"guard_type"`
	Phase     string          `json:"phase"`
	Config    json.RawMessage `json:"config"`
	Priority  int32           `json:"priority"`
	Enabled   bool            `json:"enabled"`
	CreatedAt time.Time       `json:"created_at"`
	UpdatedAt time.Time       `json:"updated_at"`
}

type Organization struct {
	ID        uuid.UUID `json:"id"`
	Name      string    `json:"name"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type Source struct {
	ID             uuid.UUID   `json:"id"`
	OrganizationID uuid.UUID   `json:"organization_id"`
	Name           string      `json:"name"`
	Route          string      `json:"route"`
	Description    pgtype.Text `json:"description"`
	LlmProvider    string      `json:"llm_provider"`
	LlmModel       string      `json:"llm_model"`
	LlmApiKeyEnc   string      `json:"llm_api_key_enc"`
	PassCriteria   string      `json:"pass_criteria"`
	PassThreshold  pgtype.Int4 `json:"pass_threshold"`
	Enabled        bool        `json:"enabled"`
	CreatedAt      time.Time   `json:"created_at"`
	UpdatedAt      time.Time   `json:"updated_at"`
}

type User struct {
	ID             uuid.UUID `json:"id"`
	OrganizationID uuid.UUID `json:"organization_id"`
	Email          string    `json:"email"`
	PasswordHash   string    `json:"password_hash"`
	CreatedAt      time.Time `json:"created_at"`
	UpdatedAt      time.Time `json:"updated_at"`
}