# openapi.yaml
openapi: "3.0.3"
info:
  title: Sheeld API
  description: Segment for LLM guardrails — a full LLM proxy that validates input, proxies LLM calls, and validates output.
  version: "1.0.0"
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8080
    description: Local development
security: []
paths:
...
  /v1/auth/api-keys:
    post:
      summary: Create an API key
      operationId: createApiKey
      tags: [Auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key:
                    $ref: "#/components/schemas/ApiKey"
                  raw_key:
                    type: string
                    description: Only returned at creation time
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      summary: List API keys
      operationId: listApiKeys
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: API key list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/api-keys/{id}:
    delete:
      summary: Revoke an API key
      operationId: revokeApiKey
      tags: [Auth]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key (shld_* prefix)
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
        created_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
          nullable: true
...